requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
python-dotenv
pytest
pytest-asyncio
httpx
python-multipart>=0.0.9
boto3
pyjwt
cryptography
